# feat(runner): add comprehensive benchmarks for cell operations