## Summary
- Add extensive benchmark suite for Cell API performance testing
- Cover all major Cell operations including schemaless and schema-based variants
- Provide baseline metrics for future performance optimizations

## Changes
- **Core operations**: Benchmarks for creation, get/set, update, and send operations
- **Schema validation**: Compare performance between schemaless and schema-based cells
- **Nested access**: Test key() navigation and deeply nested object access
- **Array operations**: Measure push() and map() performance
- **Subscriptions**: Test sink() execution with value changes
- **Complex scenarios**: Include asCell references, proxy operations, and concurrent access
- **Large data structures**: Benchmark handling of 1000-item arrays and 10-level deep objects

## Technical Details
- All benchmarks run 1000 iterations for consistency
- Use proper setup/cleanup with transaction management
- Include both simple and complex data structures
- Test schema validation overhead across different schema types
- Measure real subscription execution with runtime.idle()

## Test Coverage
The benchmark suite covers:
- Cell creation (simple, with schema, immutable)
- Read operations (get, getRaw, getAsQueryResult)
- Write operations (set, send, update, push)
- Navigation (key access for objects and arrays)
- Transformations (asSchema, withTx)
- Links and references (getAsLink, nested cells)
- Proxy operations for both reads and writes
- Concurrent operations across multiple cells

----

Current results on my machine:

```
    CPU | Apple M3 Max
Runtime | Deno 2.3.5 (aarch64-apple-darwin)

file:///Users/berni/src/labs2/packages/runner/test/cell.bench.ts

benchmark                                                   time/iter (avg)        iter/s      (min … max)           p75      p99     p995
----------------------------------------------------------- ----------------------------- --------------------- --------------------------
Cell creation - simple schemaless (1000x)                           69.1 ms          14.5 ( 64.7 ms …  75.6 ms)  70.5 ms  75.6 ms  75.6 ms
Cell creation - with JSON schema (1000x)                            91.7 ms          10.9 ( 88.2 ms …  95.1 ms)  92.6 ms  95.1 ms  95.1 ms
Cell creation - immutable (1000x)                                   27.5 ms          36.4 ( 25.3 ms …  28.8 ms)  28.2 ms  28.8 ms  28.8 ms
Cell creation - simple with schema (1000x)                          66.4 ms          15.1 ( 64.0 ms …  68.6 ms)  68.0 ms  68.6 ms  68.6 ms
Cell creation - object with nested schema (1000x)                  156.8 ms           6.4 (147.3 ms … 226.0 ms) 155.1 ms 226.0 ms 226.0 ms
Cell creation - array with schema (1000x)                          141.5 ms           7.1 (134.2 ms … 146.5 ms) 145.7 ms 146.5 ms 146.5 ms
Cell get - simple value schemaless (1000x)                          40.4 ms          24.8 ( 39.0 ms …  42.8 ms)  41.1 ms  42.8 ms  42.8 ms
Cell get - complex object schemaless (1000x)                        41.3 ms          24.2 ( 40.1 ms …  43.2 ms)  41.5 ms  43.2 ms  43.2 ms
Cell getRaw - complex object schemaless (1000x)                    587.5 µs         1,702 (400.2 µs …  59.0 ms) 462.4 µs   3.5 ms   4.1 ms
Cell get - simple value with schema (1000x)                         41.7 ms          24.0 ( 40.3 ms …  43.8 ms)  42.7 ms  43.8 ms  43.8 ms
Cell get - complex object with schema (1000x)                         1.1 s           0.9 (   1.0 s …    1.1 s)    1.1 s    1.1 s    1.1 s
Cell set - simple value schemaless (1000x)                          19.4 ms          51.6 ( 17.6 ms …  21.6 ms)  19.9 ms  21.6 ms  21.6 ms
Cell send - simple value schemaless (1000x)                         19.0 ms          52.6 ( 18.0 ms …  20.5 ms)  19.4 ms  20.5 ms  20.5 ms
Cell update - partial object update schemaless (1000x)             129.8 ms           7.7 (121.3 ms … 175.7 ms) 126.2 ms 175.7 ms 175.7 ms
Cell set - simple value with schema (1000x)                         19.2 ms          52.0 ( 17.4 ms …  21.4 ms)  19.6 ms  21.4 ms  21.4 ms
Cell update - partial object update with schema (1000x)            127.9 ms           7.8 (125.5 ms … 132.1 ms) 128.3 ms 132.1 ms 132.1 ms
Cell key - nested access schemaless (1000x)                        632.7 ms           1.6 (616.3 ms … 642.6 ms) 642.0 ms 642.6 ms 642.6 ms
Cell key - array access schemaless (1000x)                         448.6 ms           2.2 (438.9 ms … 465.9 ms) 451.6 ms 465.9 ms 465.9 ms
Cell key - nested access with schema (1000x)                       661.8 ms           1.5 (644.6 ms … 672.5 ms) 668.0 ms 672.5 ms 672.5 ms
Cell key - array access with schema (1000x)                        453.1 ms           2.2 (444.2 ms … 466.5 ms) 456.9 ms 466.5 ms 466.5 ms
Cell asSchema - schema transformation (1000x)                      332.5 ms           3.0 (318.0 ms … 343.3 ms) 340.6 ms 343.3 ms 343.3 ms
Cell withTx - transaction switching (1000x)                         72.9 ms          13.7 ( 71.4 ms …  74.7 ms)  73.5 ms  74.7 ms  74.7 ms
Cell getAsQueryResult - proxy creation schemaless (1000x)          208.6 ms           4.8 (198.5 ms … 259.1 ms) 206.1 ms 259.1 ms 259.1 ms
Cell proxy - property writes schemaless (1000x)                      3.7 ms         268.2 (  3.2 ms …   7.4 ms)   3.6 ms   5.7 ms   7.4 ms
Cell get - complex object with asCell schema (1000x)               665.9 ms           1.5 (658.0 ms … 676.9 ms) 668.8 ms 676.9 ms 676.9 ms
Cell push - array append schemaless (1000x)                        304.6 ms           3.3 (290.4 ms … 333.8 ms) 306.5 ms 333.8 ms 333.8 ms
Cell array - proxy map operation schemaless (1000x)                   7.9 s           0.1 (   7.7 s …    8.1 s)    8.0 s    8.1 s    8.1 s
Cell push - array append with schema (1000x)                       304.1 ms           3.3 (292.5 ms … 347.6 ms) 301.4 ms 347.6 ms 347.6 ms
Cell array - map operation with schema (1000x)                       15.9 s           0.1 (  15.7 s …   16.2 s)   16.0 s   16.2 s   16.2 s
Cell getAsLink - link generation schemaless (1000x)                108.7 µs         9,202 ( 91.7 µs …   3.9 ms) 103.8 µs 222.1 µs 509.8 µs
Cell getAsLink - with options (1000x)                              237.1 µs         4,218 (213.2 µs …   1.9 ms) 230.1 µs 644.2 µs   1.1 ms
Cell sink - subscription execution (1000x)                         112.1 ms           8.9 (107.8 ms … 117.2 ms) 114.6 ms 117.2 ms 117.2 ms
Cell sink - subscription execution with schema (1000x)             529.5 ms           1.9 (510.1 ms … 539.7 ms) 537.6 ms 539.7 ms 539.7 ms
Cell complex - schema with asCell references (1000x)                  1.7 s           0.6 (   1.7 s …    1.7 s)    1.7 s    1.7 s    1.7 s
Cell complex - nested cell references (1000x)                      128.8 ms           7.8 (125.3 ms … 136.4 ms) 130.4 ms 136.4 ms 136.4 ms
Cell schema - complex validation (1000x)                            32.7 ms          30.5 ( 30.1 ms …  48.1 ms)  32.6 ms  48.1 ms  48.1 ms
Cell large - array with 1000 items (1000x get)                      84.7 ms          11.8 ( 81.5 ms …  88.1 ms)  85.7 ms  88.1 ms  88.1 ms
Cell large - deeply nested object (1000x navigation)                  2.6 s           0.4 (   2.6 s …    2.7 s)    2.6 s    2.7 s    2.7 s
Cell concurrent - multiple cells (1000x)                           468.9 ms           2.1 (446.2 ms … 483.1 ms) 475.3 ms 483.1 ms 483.1 ms
Cell equals - comparison operations (1000x)                        308.3 µs         3,244 (273.1 µs …   3.4 ms) 306.2 µs 423.6 µs 517.2 µs
```
    
<!-- This is an auto-generated description by cubic. -->
---

## Summary by cubic
Added a comprehensive benchmark suite for the Cell API to measure performance across all major operations, including schemaless and schema-based variants.

- **New Features**
  - Benchmarks cover cell creation, get/set/update, nested access, array operations, schema validation, subscriptions, proxy operations, large data structures, and concurrent access.
  - Tests run 1000 iterations per operation for consistent metrics and include both simple and complex data structures.

<!-- End of auto-generated description by cubic. -->